api/requirements.txt
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
